+++ src/TranslateServiceProvider.php
@@ -3,6 +3,7 @@
 namespace Weglot\Translate;
 
 use Illuminate\Support\ServiceProvider;
+use Weglot\Translate\Commands\CacheClearCommand;
 use Weglot\Translate\Providers\BladeServiceProvider;
 use Weglot\Translate\Providers\RouterServiceProvider;
 
@@ -17,7 +18,7 @@ class TranslateServiceProvider extends ServiceProvider
      *
      * @var string
      */
-    const VERSION = '0.2.1';
+    const VERSION = '0.3.0';
 
     /**
      * Bootstrap services.
@@ -43,6 +44,12 @@ class TranslateServiceProvider extends ServiceProvider
 
         $this->app->register(BladeServiceProvider::class);
         $this->app->register(RouterServiceProvider::class);
+
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                CacheClearCommand::class
+            ]);
+        }
     }
 
     /**
@@ -52,6 +59,6 @@ class TranslateServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        //
+        $this->app->bind('weglot.cache', 'Weglot\\Translate\\Cache\\Cache');
     }
 }